# feat(java): scoped meta share mode for type forward/backward compaibility