--- a/java/fury-core/src/main/java/org/apache/fury/config/FuryBuilder.java
+++ b/java/fury-core/src/main/java/org/apache/fury/config/FuryBuilder.java
@@ -245,6 +245,9 @@ public final class FuryBuilder {
    */
   public FuryBuilder withScopedMetaShare(boolean scoped) {
     scopedMetaShareEnabled = scoped;
+    if (scoped) {
+      metaShareEnabled = true;
+    }
     return this;
   }