--- a/java/fury-core/src/main/java/org/apache/fury/meta/ClassDef.java
+++ b/java/fury-core/src/main/java/org/apache/fury/meta/ClassDef.java
@@ -767,7 +767,7 @@ public class ClassDef implements Serializable {
     return buildClassDef(classResolver, type, fields, new byte[0]);
   }
 
-  private static ClassDef buildClassDef(
+  public static ClassDef buildClassDef(
       ClassResolver classResolver, Class<?> type, List<Field> fields, byte[] extMeta) {
     return ClassDefEncoder.buildClassDef(classResolver, type, fields, extMeta);
   }