--- a/java/fury-core/src/main/java/org/apache/fury/meta/ClassSpec.java
+++ b/java/fury-core/src/main/java/org/apache/fury/meta/ClassSpec.java
@@ -37,8 +37,12 @@ public class ClassSpec {
 
   @Override
   public boolean equals(Object o) {
-    if (this == o) return true;
-    if (o == null || getClass() != o.getClass()) return false;
+    if (this == o) {
+      return true;
+    }
+    if (o == null || getClass() != o.getClass()) {
+      return false;
+    }
     ClassSpec classSpec = (ClassSpec) o;
     return entireClassName.equals(classSpec.entireClassName);
   }