… integration_tests/latest_jdk_tests/src/test/java/org/apache/fury/integration_tests/RecordSerializersTest.java
… java/benchmark/pom.xml
… java/benchmark/src/main/java/org/apache/fury/benchmark/UserTypeDeserializeSuite.java
… java/benchmark/src/main/java/org/apache/fury/benchmark/UserTypeSerializeSuite.java
… java/benchmark/src/main/java/org/apache/fury/benchmark/state/FuryState.java
… java/benchmark/src/main/java/org/apache/fury/benchmark/state/KryoState.java
… java/fury-core/src/main/java/org/apache/fury/meta/Encoders.java
… java/fury-core/src/main/java/org/apache/fury/reflect/ReflectionUtils.java
… java/fury-core/src/main/java/org/apache/fury/resolver/ClassInfo.java
… java/fury-core/src/main/java/org/apache/fury/resolver/ClassResolver.java
… java/fury-core/src/main/java/org/apache/fury/resolver/SerializationContext.java
… java/fury-core/src/main/java/org/apache/fury/serializer/ArraySerializers.java
… java/fury-core/src/main/java/org/apache/fury/serializer/MetaSharedSerializer.java
… java/fury-core/src/main/java/org/apache/fury/serializer/NonexistentClass.java
… java/fury-core/src/main/java/org/apache/fury/serializer/NonexistentClassSerializers.java
… java/fury-core/src/main/java/org/apache/fury/serializer/ObjectSerializer.java
… java/fury-core/src/main/java/org/apache/fury/serializer/collection/AbstractCollectionSerializer.java
… java/fury-core/src/test/java/org/apache/fury/CyclicTest.java
… java/fury-core/src/test/java/org/apache/fury/FuryInitPerf.java
… java/fury-core/src/test/java/org/apache/fury/ThreadSafeFuryTest.java
… java/fury-core/src/test/java/org/apache/fury/meta/ClassDefEncoderTest.java
… java/fury-core/src/test/java/org/apache/fury/meta/ClassDefTest.java
… java/fury-core/src/test/java/org/apache/fury/reflect/ReflectionUtilsTest.java
… java/fury-core/src/test/java/org/apache/fury/resolver/MetaContextTest.java
… java/fury-core/src/test/java/org/apache/fury/resolver/SerializationContextTest.java
… java/fury-core/src/test/java/org/apache/fury/serializer/MetaSharedCompatibleTest.java
… java/fury-core/src/test/java/org/apache/fury/serializer/NonexistentClassSerializersTest.java
… java/fury-core/src/test/java/org/apache/fury/serializer/ProtocolInteroperabilityTest.java
… java/fury-core/src/test/java/org/apache/fury/serializer/collection/ChildContainerSerializersTest.java