@chaokunyang 2024-05-29
Hi @MrChang0, this PR should make the meta share mode replace kv compatible mode for better performance and smaller space cost