@MrChang0 2024-05-30
It fine for me, I will try to experiment this feature.
btw, how about increase that init buffer size in `org.apache.fury.meta.ClassDefEncoder#encodeClassDef`. take less growth operate may take more performance.