@chaokunyang 2024-05-30
> encodeClassDef

Would you like to submit a PR?