@chaokunyang 2024-05-30
Struct perf:
```
Before:
fury_deserialize                    array        STRUCT         false  thrpt   10  4103952.355 ± 461542.704  ops/s
fury_deserialize_compatible         array        STRUCT         false  thrpt   10  2104530.838 ± 283950.067  ops/s

After:
Benchmark                                             (bufferType)  (objectType)  (references)   Mode  Cnt        Score        Error  Units
fury_deserialize                    array        STRUCT         false  thrpt   10  3846945.672 ± 281139.320  ops/s
fury_deserialize_compatible         array        STRUCT         false  thrpt   10  3119394.319 ± 476422.038  ops/s
```

Size:
```
Before:
2024-05-30 02:06:42 INFO  FuryState:157 [fury_deserialize-jmh-worker-1] - ======> Fury | STRUCT | false | array | 710 |
2024-05-30 02:07:19 INFO  FuryState:157 [fury_deserialize_compatible-jmh-worker-1] - ======> Fury | STRUCT | false | array | 1134 |

After:
2024-05-30 02:01:16 INFO  FuryState:157 [fury_deserialize-jmh-worker-1] - ======> Fury | STRUCT | false | array | 710 |
2024-05-30 02:01:52 INFO  FuryState:157 [fury_deserialize_compatible-jmh-worker-1] - ======> Fury | STRUCT | false | array | 1227 |
```

Size bloat a little, we can optimize it later